feat(servers): use the server a tier below if it does not exist
If a server does not exist in a tier, use the tier below it.
For example, if a user is in the test access level, but a server for that game only exists on prod, use that server.
Same for the dev access level, but it goes dev --> test --> prod