fix: Seperate token checks by type
Give each expected token type a seperate function and un-export the "generic" one.

This required deciding what the "expected" token type is for all endpoints. Possibly contraversial choices are:

- API tokens only on the HTTP API (no service tokens)
- API tokens only on the gRPC APIs
- API tokens only on ExchangeTokenForUserData (I could only find 1 user of this API - BOSS - and they use API tokens here)

Notably, there is now no way to get the account server to validate a service token - services must decrypt them locally. We should probably add an API for this for the benefit of third parties.