feat(nasc): Add enforcement of custom patches to required titles
If the supplied game server ID doesn't match the one present on the
server database, then assume the title requires patches on the user's
end. Give this behavior a custom error code 002-0152 to facilitate
support to unaware users.

Also drop the legacy fallback of the friends server to the test
environment, since the server should be available to anyone anyway, and
we already have the production config for it.

Alongside that, update the maintenance error code to use the proper
002-0101 as expected.